+++ src/main/java/utec/proyectofinal/Proyecto/Final/UTEC/dtos/response/ClienteDTO.java
@@ -4,7 +4,7 @@ import lombok.Data;
 
 @Data
 public class ClienteDTO {
-    private Integer clienteID;
+    private Long clienteID;
     private String nombre;
     private String contacto;
     private String email;
